requirements.txt
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
pyenchant==3.2.2
scikit_learn==1.3.2
scipy==1.11.4
tensorflow==2.15.0
tensorflow_intel==2.15.0
wordninja==2.0.0
langdetect==1.0.9